[modules/pin_cfg] move uC specific code to port/ucontroller/
