+++ modules/CMakeLists.txt
@@ -4,7 +4,6 @@ set(PROJ_HDRS ${PROJ_HDRS} ${MODULE_PATH} )
 
 set(PROJ_SRCS ${PROJ_SRCS} ${CMAKE_CURRENT_SOURCE_DIR}/main.c)
 set(PROJ_SRCS ${PROJ_SRCS} ${MODULE_PATH}/utils.c)
-set(PROJ_SRCS ${PROJ_SRCS} ${MODULE_PATH}/pin_cfg.c)
 set(PROJ_SRCS ${PROJ_SRCS} ${MODULE_PATH}/i2c.c)
 set(PROJ_SRCS ${PROJ_SRCS} ${MODULE_PATH}/led.c)
 set(PROJ_SRCS ${PROJ_SRCS} ${MODULE_PATH}/ipmb.c ${MODULE_PATH}/ipmi.c)
